chore(backend): get_settings function + unit tests
Previously it was just a variable called on module init. But that is a
problem if I want to unit test the settings class it  would error if it
didn't have the right ENV vars set. So now it is now a "singleton"
function.

Also added unit tests that test config + websockets!